factor TlsConfiguration parser from extractor
